vm-api/pyproject.toml
[project]
name = "vm-api"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "asyncpg>=0.29",
    "httpx>=0.27",
    "pydantic[email]>=2.0",
    "reportlab>=4.0",
    "markdown>=3.6",
    "python-dotenv>=1.0",
    "filelock>=3.13",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
